ocr-backend/requirements.txt
paddlepaddle==2.6.1
paddleocr==2.7.3
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.3.0
python-dotenv==1.0.1
